fixed norm error, add exec_builtin.c and exec_external.c
